+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Update ESLint to 7.30.0
+- Changed array-bracket-newline and array-element-newline to 'consistent'
 
 ## [1.0.0-alpha.2] - 2021-07-01